backend/requirements.txt
Flask==3.0.0
Flask-Session==0.6.0
python-dotenv==1.0.0
qrcode[pil]==7.4.2
Pillow==10.2.0
gunicorn==21.2.0
google-generativeai==0.5.4
